major refactor
= avoid scalars for now, as these are not validated anyway yet by graphql-js for directive arguments
= pare down initial list of supported arguments for merge, see below
= move code into new package
= spit code/tests into multiple files
= start adding functionality/tests for subschemaConfig transformer

initial arguments will be nothing, which will use all selectionSets passed to the first argument a la _entities-type approach OR argsExpr, soon to be followed by key/keyField/args options
