--- packages/schema/src/schemaTransforms/index.ts
@@ -1,1 +1,0 @@
-export * from './stitchingDirectives';
