+++ packages/type-merging-directives/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@graphql-tools/type-merging-directives",
+  "version": "6.0.0",
+  "description": "A set of utils for faster development of GraphQL tools",
+  "repository": "git@github.com:ardatan/graphql-tools.git",
+  "license": "MIT",
+  "sideEffects": false,
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.esm.js",
+  "typings": "dist/index.d.ts",
+  "typescript": {
+    "definition": "dist/index.d.ts"
+  },
+  "peerDependencies": {
+    "graphql": "^14.0.0 || ^15.0.0"
+  },
+  "buildOptions": {
+    "input": "./src/index.ts"
+  },
+  "dependencies": {
+    "@graphql-tools/delegate": "^7.0.0",
+    "@graphql-tools/utils": "^7.0.0",
+    "tslib": "~2.0.1"
+  },
+  "devDependencies": {
+    "@graphql-tools/schema": "^7.0.0"
+  },
+  "publishConfig": {
+    "access": "public",
+    "directory": "dist"
+  }
+}
