+++ packages/type-merging-directives/src/defaultTypeMergingDirectiveOptions.ts
@@ -1,0 +1,5 @@
+export const defaultTypeMergingDirectiveOptions = {
+  baseDirectiveName: 'base',
+  computedDirectiveName: 'computed',
+  mergeDirectiveName: 'merge',
+};
