+++ packages/type-merging-directives/src/index.ts
@@ -1,0 +1,2 @@
+export * from './typeMergingDirectives';
+export * from './types';
