+++ packages/utils/src/index.ts
@@ -49,4 +49,3 @@ export * from './getArgumentValues';
 export * from './valueMatchesCriteria';
 export * from './isAsyncIterable';
 export * from './isDocumentNode';
-export { parseMergeArgsExpr } from './parseMergeArgsExpr';
