+++ packages/utils/src/types.ts
@@ -36,22 +36,3 @@ export interface PruneSchemaOptions {
    */
   skipUnusedTypesPruning?: boolean;
 }
-
-export interface ParsedMergeArgsExpr {
-  args: Record<string, any>;
-  keyDeclarations: Array<KeyDeclaration>;
-  expansions: Array<Expansion>;
-}
-
-export interface KeyDeclaration {
-  valuePath: Array<string | number>;
-  keyPath: Array<string>;
-}
-
-export interface Expansion {
-  valuePath: Array<string | number>;
-  value: any;
-  keyDeclarations: Array<KeyDeclaration>;
-}
-
-export type VariablePaths = Record<string, Array<string | number>>;
